Crate roles_types

Structs

AssignRoleInput

Input structure for assigning roles

AssigneeRoleClaimLinkTag

PendingUnassignmentLinkTag

Input structure for unassigning roles

Properties

App properties (progenitor info)

RequestUnassignRoleInput

Input structure for unassigning roles

Enums

AssigneeRoleClaimLinkMode

Constants

ADMIN_ROLE

Hardcoded admin role need for creating other roles

ROLE_CLAIM_LINK_TYPE_INDEX

Functions

validate_agent_had_undeleted_role_claim_at_the_time

Validate that agents had acces to the role at the time of the action (an undeleted claim earlier in source chain)

validate_agent_had_undeleted_role_claim_at_the_time_with_zome_index

Validate that agents had acces to the role at the time (with zome index)